+++ controllers/project-controllers.ts
@@ -1,7 +1,8 @@
 import { Request, Response, NextFunction } from "express";
 import projectService from "../services/project-service";
 import { updateProject,createProject } from "../schemas/project-schema";
-
+import {v2 as cloudinary } from 'cloudinary'
+import { extractPublicId } from "cloudinary-build-url";
 
 class projectControllers {
   async getAllProject(req: Request, res: Response, next: NextFunction) {
@@ -35,6 +36,9 @@ class projectControllers {
       const body = req.body
       const validateProject = await updateProject.validateAsync(body)
       const update = await projectService.updateProject(id, validateProject);
+      const image = update.image
+      const publicId = extractPublicId(image)
+      cloudinary.uploader.destroy(publicId)
       res.json({data:update , message:"edit success"});
     } catch (error) {
       next(error);
@@ -44,6 +48,16 @@ class projectControllers {
   async deleteProject(req:Request,res:Response,next:NextFunction){
     try{
       const {id} = req.params
+      const project = await projectService.getProjectById(id)
+      const image = project?.image
+
+      if(!image){
+        res.status(400).json({message:"image not found"})
+        return
+      }
+
+      const publicId = extractPublicId(image)
+      await cloudinary.uploader.destroy(publicId)
       await projectService.deleteProject(id)
       res.json({message:"deleted"})
     }catch(error){
